refactor: modify class name and constructor
